Resolve circular dependency between traits and ir modules
